fix(consensus): widen cumulative supply accumulators to u128
Phase-1 emission (~1.22e21 picocredits ≈ 1.22B BTH) exceeds u64::MAX
(~1.84e19 pico ≈ 18.4M BTH). With overflow-checks=false in release, the
u64 supply accumulators wrapped silently once cumulative emission passed
~18.4M BTH. Because total_mined feeds calculate_block_reward(), a wrapped
value corrupted the emission schedule itself and diverged nodes.

Widen the cumulative monetary accumulators to u128:
- ChainState.total_mined / total_fees_burned
- EmissionController.total_emitted / total_burned (in-memory mirror
  restored from chain state via from_chain_state)
- MintingWork.total_minted (passed to calculate_block_reward)
- calculate_block_reward's total_supply param, plus a u128 reimplementation
  of the tail-reward formula (kept bit-for-bit identical to the cluster-tax
  u64 version) so Phase-2 rewards are correct at real picocredit supply.

Persistence: META_TOTAL_MINED / META_FEES_BURNED now serialize as 16-byte
LE (was 8). This is a consensus-state format change that rides the planned
testnet reset (#323); no migration path is needed.

RPC: handle_chain_info / handle_supply_info now emit totalMined,
totalFeesBurned and circulatingSupply as decimal strings (u128 always
exceeds JS's 2^53 safe-integer limit).

Amounts stay u64: TxOutput.amount, MintingTx.reward and fee are unchanged,
so block/tx/gossip wire format is byte-for-byte identical (asserted).

Adds regression tests: accumulation past u64::MAX (debug+release),
correct calculate_block_reward beyond 2^64 pico, 16-byte metadata
persist/reload, ChainState snapshot serde round-trip > u64::MAX, and a
wire-format guard that monetary fields remain u64-sized.

Closes #333

+++ botho/src/consensus/lottery.rs
@@ -1174,7 +1174,7 @@ mod tests {
             assert!(policy.lottery_emission_share(h, r) <= r);
             // Tail reward + block reward must not panic for extreme supply.
             let _ = policy.calculate_tail_reward(r);
-            let _ = crate::block::calculate_block_reward(h, r);
+            let _ = crate::block::calculate_block_reward(h, r as u128);
         }
     }
 }
